fix de entidades y controllers

+++ src/main/java/com/reba/api/person/exception/PersonNotFoundException.java
@@ -3,7 +3,7 @@ package com.reba.api.person.exception;
 import javax.persistence.EntityNotFoundException;
 
 public class PersonNotFoundException extends EntityNotFoundException {
-    public PersonNotFoundException(Long personId) {
+    public PersonNotFoundException(Integer personId) {
         super(String.format("Persona con id %d no encontrada", personId));
     }
